chore: DB 기술 분리 #144
